+++ next.config.ts
@@ -4,6 +4,7 @@ import { initOpenNextCloudflareForDev } from "@opennextjs/cloudflare";
 initOpenNextCloudflareForDev();
 
 const nextConfig: NextConfig = {
+  reactStrictMode: false,
   images: {
     unoptimized: true,
   },